requirements.txt
PyPDF2==1.26.0
fpdf2==2.4.2